+++ src/collective/googleanalytics/tests/base.py
@@ -1,44 +1,17 @@
+from collective.googleanalytics.testing import INTEGRATION_TESTING
+from collective.googleanalytics.testing import FUNCTIONAL_TESTING
 
-from Products.Five import zcml
-from Products.Five import fiveconfigure
-from Products.PloneTestCase import PloneTestCase as ptc
-from Products.PloneTestCase.layer import onsetup
-from Testing import ZopeTestCase as ztc
+import unittest
 
 
-@onsetup
-def setup_product():
-    """Set up additional products and ZCML required to test this product.
-
-    The @onsetup decorator causes the execution of this body to be deferred
-    until the setup of the Plone site testing layer.
-    """
-
-    # Load the ZCML configuration for this package and its dependencies
-
-    fiveconfigure.debug_mode = True
-    import collective.googleanalytics
-    zcml.load_config('configure.zcml', collective.googleanalytics)
-    fiveconfigure.debug_mode = False
-
-    # We need to tell the testing framework that these products
-    # should be available. This can't happen until after we have loaded
-    # the ZCML.
-
-    ztc.installPackage('collective.googleanalytics')
-
-# The order here is important: We first call the deferred function and then
-# let PloneTestCase install it during Plone site setup
-
-setup_product()
-ptc.setupPloneSite(products=['collective.googleanalytics'])
-
-
-class TestCase(ptc.PloneTestCase):
+class TestCase(unittest.TestCase):
     """Base class used for test cases
     """
 
+    layer = INTEGRATION_TESTING
 
-class FunctionalTestCase(ptc.FunctionalTestCase):
+class FunctionalTestCase(unittest.TestCase):
     """Test case class used for functional (doc-)tests
     """
+
+    layer = FUNCTIONAL_TESTING